more predictability
not 100% determinism, because some indeterminism
seems to arise from the same issue inside JastaddJ
(or even Jastadd-no-J)

the issue is HashMap/HashSet<Object ...> that is,
Hash* types with object identity used as comparison
and POINTER values used for the hash. these differ
per run and cause indeterminism.

Jastadd "contributors" seem to suffer from this,
grep for IdentityHashSet.