Fix some code that exploited a bug in JastAdd
This is a bit ugly, but the original implementation used the fact that you could
omit the return value from an inherited attribute. The possibility to omit the
return value was considered a bug in JastAdd, and it has been fixed in newer
versions. This, in turn, breaks the PT compiler when upgrading to a newer
version of JastAddJ.

See http://svn.cs.lth.se/trac/jastadd-trac/ticket/70 for the JastAdd bug report.